# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: C5 360相机巡查标注系统 API
  version: 1.0.0
  license:
    name: Proprietary
  description: >
    Spec-first OpenAPI 3 contract for project C5 (360-camera inspection &
    annotation system). This single document is the authoritative source for
    oapi-codegen v2 → Go server stubs + TypeScript + Kotlin/Retrofit clients.


    Conventions:

    - Every response uses the shared envelope `{success, data, error, meta?}`.
      List endpoints populate `meta` with `{total, page, page_size}`.

    - All IDs are int64 (BIGSERIAL in the DB). `client_uuid` / `job_uuid` are
      UUID strings (offline idempotency / public job handles).

    - All geometry on the wire is WGS84 / SRID 4326 GeoJSON. GCJ-02 is NEVER
      persisted or transmitted; it is a client-only render-boundary concern.

    - All timestamps are RFC3339 / ISO-8601 UTC (`format: date-time`).

    - Error codes are a frozen, canonical enum (see ErrorObject.code).


    Out-of-band (NOT expressible in OpenAPI):

    - `GET /exports/{job_uuid}/events` is a Server-Sent Events (`text/event-stream`)
      endpoint emitting `event: progress` and a terminal `event: done`. It is
      hand-written with `GET /exports/{job_uuid}` as the polling fallback
      (Tencent CLB/CDN may buffer SSE). It is intentionally absent from `paths`.

    - Dictionary / config pull uses ETag (`If-None-Match: "<content_hash>"` →
      304 Not Modified when unchanged). 304 responses carry no body.

...
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /inspections:
    get:
      tags: [inspections]
      operationId: listInspections
      summary: List inspection sessions
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: project_id
          in: query
          schema: { type: integer, format: int64 }
        - name: inspector_id
          in: query
          schema: { type: integer, format: int64 }
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/InspectionStatus'
        - $ref: '#/components/parameters/FromParam'
        - $ref: '#/components/parameters/ToParam'
      responses:
        '200':
          description: Inspection list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InspectionListResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /inspections/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [inspections]
      operationId: getInspection
      summary: Get an inspection session by id
      responses:
        '200':
          description: Inspection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InspectionResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /inspections/start:
    post:
      tags: [inspections]
      operationId: startInspection
      summary: Start an inspection session (status → IN_PROGRESS)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InspectionStartRequest'
      responses:
        '201':
          description: Inspection started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InspectionResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '500':
          $ref: '#/components/responses/InternalError'
  /inspections/{id}/finish:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    post:
      tags: [inspections]
      operationId: finishInspection
      summary: Finish an inspection (computes mileage_meters + duration_seconds, status → FINISHED)
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InspectionFinishRequest'
      responses:
        '200':
          description: Inspection finished
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InspectionResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '500':
          $ref: '#/components/responses/InternalError'
  /inspections/{id}/trajectory:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [inspections]
      operationId: getInspectionTrajectory
      summary: Get an inspection's route + raw points (WGS84 GeoJSON)
      responses:
        '200':
          description: Trajectory (route LineString + ordered points)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrajectoryResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # ===========================================================================
  # Problems
  # ===========================================================================
...
  /exports:
    post:
      tags: [exports]
      operationId: createExport
      summary: Enqueue an async Excel export job
      description: >
        Returns `201 {job_uuid, status:"PENDING"}`. Track progress via SSE at
        `GET /exports/{job_uuid}/events` (out-of-band, not in this spec) or poll
        `GET /exports/{job_uuid}`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportJobCreate'
      responses:
        '201':
          description: Export job enqueued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportJobResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '500':
          $ref: '#/components/responses/InternalError'
  /exports/{job_uuid}:
    parameters:
      - $ref: '#/components/parameters/JobUuidPath'
    get:
      tags: [exports]
      operationId: getExport
      summary: Poll an export job's status (SSE polling fallback)
      description: >
        On SUCCEEDED, `result_url` is a signed CDN URL to the produced `.xlsx`.
        Clients poll every ~2s if the SSE stream drops.
      responses:
        '200':
          description: Export job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportJobResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # ===========================================================================
  # Offline sync
  # ===========================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived access JWT (golang-jwt/v5). Refresh is an opaque Redis token.
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
...
    JobUuidPath:
      name: job_uuid
      in: path
      required: true
      schema: { type: string, format: uuid }
    PageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSizeParam:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 20 }
    FromParam:
      name: from
      in: query
      description: Inclusive lower time bound (RFC3339 UTC)
      schema: { type: string, format: date-time }
    ToParam:
      name: to
      in: query
      description: Exclusive upper time bound (RFC3339 UTC)
      schema: { type: string, format: date-time }
...
  responses:
...
    Unauthenticated:
      description: Missing/invalid/expired bearer token (UNAUTHENTICATED | TOKEN_EXPIRED)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: casbin (sub,obj,act) denied (FORBIDDEN)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found or soft-deleted (NOT_FOUND)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Unique conflict / RESTRICT / illegal state transition (CONFLICT)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ValidationFailed:
      description: Boundary schema / business validation failed (VALIDATION_FAILED)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalError:
      description: Unhandled error (INTERNAL); never leaks stack/SQL
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    Meta:
      type: object
      properties:
        total: { type: integer, format: int64 }
        page: { type: integer }
        page_size: { type: integer }
    ErrorCode:
      type: string
      enum:
        - VALIDATION_FAILED
        - UNAUTHENTICATED
        - TOKEN_EXPIRED
        - FORBIDDEN
        - NOT_FOUND
        - CONFLICT
        - MEDIA_VERIFY_FAILED
        - DICT_VERSION_RETIRED
        - RATE_LIMITED
        - INTERNAL
    ErrorDetail:
      type: object
      required: [field, code, message]
      properties:
        field: { type: string }
        code: { type: string }
        message: { type: string }
    ErrorObject:
      type: object
      required: [code, message]
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message: { type: string }
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
    ErrorResponse:
      type: object
      required: [success, data, error]
      properties:
        success: { type: boolean, example: false }
        data:
          type: object
          nullable: true
          example: null
        error:
          $ref: '#/components/schemas/ErrorObject'
        meta:
          $ref: '#/components/schemas/Meta'
...
    GeoJSONPoint:
      type: object
      required: [type, coordinates]
      properties:
        type:
          type: string
          enum: [Point]
        coordinates:
          type: array
          description: '[longitude, latitude] in WGS84'
          items: { type: number }
          minItems: 2
          maxItems: 2
    GeoJSONLineString:
      type: object
      required: [type, coordinates]
      properties:
        type:
          type: string
          enum: [LineString]
        coordinates:
          type: array
          items:
            type: array
            items: { type: number }
            minItems: 2
            maxItems: 2
...
    InspectionStatus:
      type: string
      enum: [IN_PROGRESS, FINISHED, ABANDONED]
...
    JobStatus:
      type: string
      description: CANCELLED is the optional 5th value (cancellation is optional to expose).
      enum: [PENDING, RUNNING, SUCCEEDED, FAILED, CANCELLED]
    ExportType:
      type: string
      enum: [INSPECTION_RECORDS, PROBLEM_LIST, PROJECT_STATS]
...
    Inspection:
      type: object
      required:
        - id
        - client_uuid
        - project_id
        - inspector_id
        - status
        - started_at
        - duration_seconds
        - mileage_meters
        - point_count
        - created_at
        - updated_at
      properties:
        id: { type: integer, format: int64 }
        client_uuid: { type: string, format: uuid }
        project_id: { type: integer, format: int64 }
        task_id: { type: integer, format: int64, nullable: true }
        inspector_id: { type: integer, format: int64 }
        status:
          $ref: '#/components/schemas/InspectionStatus'
        started_at: { type: string, format: date-time }
        ended_at: { type: string, format: date-time, nullable: true }
        duration_seconds: { type: integer, format: int64 }
        mileage_meters:
          type: number
          format: double
          description: SUM via ST_Length(route_geom::geography) — meters.
        point_count: { type: integer }
        route_geom:
          $ref: '#/components/schemas/GeoJSONLineString'
        device_info:
          type: object
          additionalProperties: true
        note: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    TrajectoryPoint:
      type: object
      required: [id, client_uuid, inspection_id, seq, geom, recorded_at]
      properties:
        id: { type: integer, format: int64 }
        client_uuid: { type: string, format: uuid }
        inspection_id: { type: integer, format: int64 }
        seq: { type: integer }
        geom:
          $ref: '#/components/schemas/GeoJSONPoint'
        recorded_at: { type: string, format: date-time }
        speed: { type: number, format: float, nullable: true }
        bearing: { type: number, format: float, nullable: true }
        altitude: { type: number, format: float, nullable: true }
        accuracy: { type: number, format: float, nullable: true }
...
    ExportJob:
      type: object
      required: [id, job_uuid, type, status, progress, processed_rows, created_at]
      properties:
        id: { type: integer, format: int64 }
        job_uuid: { type: string, format: uuid }
        type:
          $ref: '#/components/schemas/ExportType'
        params:
          type: object
          additionalProperties: true
        status:
          $ref: '#/components/schemas/JobStatus'
        progress: { type: integer, minimum: 0, maximum: 100 }
        total_rows: { type: integer, nullable: true }
        processed_rows: { type: integer }
        result_cos_key: { type: string, nullable: true }
        result_bucket: { type: string, nullable: true }
        result_url:
          type: string
          nullable: true
          description: Signed CDN URL to the produced .xlsx (set on SUCCEEDED).
        error: { type: string, nullable: true }
        requested_by: { type: integer, format: int64, nullable: true }
        started_at: { type: string, format: date-time, nullable: true }
        finished_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
...
    Trajectory:
      type: object
      required: [inspection_id, points]
      properties:
        inspection_id: { type: integer, format: int64 }
        route:
          $ref: '#/components/schemas/GeoJSONLineString'
        points:
          type: array
          items:
            $ref: '#/components/schemas/TrajectoryPoint'
...
    InspectionStartRequest:
      type: object
      required: [project_id, started_at]
      properties:
        client_uuid:
          type: string
          format: uuid
          description: Offline idempotency key (server generates if omitted).
        project_id: { type: integer, format: int64 }
        task_id: { type: integer, format: int64, nullable: true }
        inspector_id:
          type: integer
          format: int64
          nullable: true
          description: Defaults to the authenticated user if omitted.
        started_at: { type: string, format: date-time }
        device_info:
          type: object
          additionalProperties: true
        note: { type: string }
    InspectionFinishRequest:
      type: object
      properties:
        ended_at:
          type: string
          format: date-time
          description: Defaults to server time if omitted. Must be >= started_at.
        status:
          $ref: '#/components/schemas/InspectionStatus'
        note: { type: string }
...
    ExportJobCreate:
      type: object
      required: [type]
      properties:
        type:
          $ref: '#/components/schemas/ExportType'
        params:
          type: object
          additionalProperties: true
          description: >
            Filters by export type — INSPECTION_RECORDS:
            {project_id, inspector_id, from, to, status};
            PROBLEM_LIST:
            {project_id, type, status, category, from, to, inspector_id, inspection_id};
            PROJECT_STATS: {project_id?, from, to, inspector_id?}.
...
    InspectionResponse:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, example: true }
        data:
          $ref: '#/components/schemas/Inspection'
        error:
          $ref: '#/components/schemas/ErrorObject'
        meta:
          $ref: '#/components/schemas/Meta'
    InspectionListResponse:
      type: object
      required: [success, data, meta]
      properties:
        success: { type: boolean, example: true }
        data:
          type: array
          items:
            $ref: '#/components/schemas/Inspection'
        error:
          $ref: '#/components/schemas/ErrorObject'
        meta:
          $ref: '#/components/schemas/Meta'
    TrajectoryResponse:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, example: true }
        data:
          $ref: '#/components/schemas/Trajectory'
        error:
          $ref: '#/components/schemas/ErrorObject'
        meta:
          $ref: '#/components/schemas/Meta'
...
    ExportJobResponse:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, example: true }
        data:
          $ref: '#/components/schemas/ExportJob'
        error:
          $ref: '#/components/schemas/ErrorObject'
        meta:
          $ref: '#/components/schemas/Meta'
...